db-populator/requirements.txt
faker
SQLAlchemy
psycopg2-binary
python-dotenv
